# Reduce implicit array allocations on caller side of method calling