This eliminates array allocations for the following common cases:

```ruby
f(1, *a)
f(1, *a, &lvar)
f(*a, **lvar)
f(*a, **lvar, &lvar)
```

This also eliminates array allocations for these less common cases that I think are still worth optimizing:

```ruby
f(1, *a, &@ivar)
f(*a, **@ivar)
f(*a, **@ivar, &@ivar)
f(*a, kw: 1)
f(*a, kw:1, &lvar)
f(*a, kw:1, &@ivar)
```

This is handled via the peephole optimizer.

In terms of safety, currently, `f(*a, &lvar)` and `f(*a, &@ivar)` both avoid array allocations, and all of the above are as safe as those in terms of safety.