@jeremyevans 2023-11-18
The reason I limited this to lvar and ivar is that those are fairly easy to handle, not risky (only pathological code fails), and very common.  The other common case is splatting of method calls, but handling that is more complex, more risky (actual code might fail), and slightly less common, so I don't think we should do that.   This could easily be extended to gvar and cvar, but those very uncommon, so I don't think it is worth it.