@jeremyevans 2023-11-18
To check how common these types of methods are, I added some metrics to the code, to see how many methods would be optimized by removing allocations:

For `ruby -e ''`  (just loads rubygems):

```
  5 : f(1, *a)
  2 : f(1, *a, &lvar) | f(1, *a, &@ivar)
```

Current stdlib:

```
139 : f(1, *a)
  3 : f(1, *a, &lvar) | f(1, *a, &@ivar)
  4 : f(*a, **lvar) | f(*a, **@ivar)
```

minitest 5.20 (bundled gem):

```
  4 : f(1, *a)
  1 : f(1, *a, &lvar) | f(1, *a, &@ivar)
  2 : f(*a, **lvar) | f(*a, **@ivar)
  2 : f(*a, **lvar, &lvar) | f(*a, **@ivar, &@ivar)
```

Rails master:

```
 77 : f(1, *a)
  5 : f(1, *a, &lvar) | f(1, *a, &@ivar)
 26 : f(*a, **lvar) | f(*a, **@ivar)
  5 : f(*a, kw: 1)
```

So all optimizations seem to be used by standard lib and common gems, except maybe `f(*a, kw:1, &lvar) | f(*a, kw:1, &@ivar)`.  I could remove that if we don't think it is worth keeping.

Code to enable metrics:

```diff
diff --git a/compile.c b/compile.c
index 41008f3cfd..88b281eedd 100644
--- a/compile.c
+++ b/compile.c
@@ -3848,6 +3848,7 @@ iseq_peephole_optimize(rb_iseq_t *iseq, LINK_ELEMENT *list, const int do_tailcal
             niobj = niobj->next;
             unsigned int flag = vm_ci_flag((const struct rb_callinfo *)OPERAND_AT(niobj, 0));
             if ((flag & VM_CALL_ARGS_SPLAT) && !(flag & (VM_CALL_KW_SPLAT|VM_CALL_ARGS_BLOCKARG))) {
+write(1, "1", 1);
                 OPERAND_AT(iobj, 0) = Qfalse;
             }
         } else if (IS_NEXT_INSN_ID(niobj, getlocal) || IS_NEXT_INSN_ID(niobj, getinstancevariable)) {
@@ -3870,6 +3871,7 @@ iseq_peephole_optimize(rb_iseq_t *iseq, LINK_ELEMENT *list, const int do_tailcal
                     *  send
                     */
                     if ((flag & VM_CALL_ARGS_BLOCKARG) && !(flag & VM_CALL_KW_SPLAT)) {
+write(1, "2", 1);
                         OPERAND_AT(iobj, 0) = Qfalse;
                     }
 
@@ -3885,6 +3887,7 @@ iseq_peephole_optimize(rb_iseq_t *iseq, LINK_ELEMENT *list, const int do_tailcal
                     *  send
                     */
                     else if (!(flag & VM_CALL_ARGS_BLOCKARG) && (flag & VM_CALL_KW_SPLAT)) {
+write(1, "3", 1);
                         OPERAND_AT(iobj, 0) = Qfalse;
                     }
                 }
@@ -3909,6 +3912,7 @@ iseq_peephole_optimize(rb_iseq_t *iseq, LINK_ELEMENT *list, const int do_tailcal
                     unsigned int flag = vm_ci_flag((const struct rb_callinfo *)OPERAND_AT(niobj, 0));
 
                     if ((flag & VM_CALL_ARGS_SPLAT) && (flag & VM_CALL_KW_SPLAT) && (flag & VM_CALL_ARGS_BLOCKARG)) {
+write(1, "4", 1);
                         OPERAND_AT(iobj, 0) = Qfalse;
                     }
                 }
@@ -3933,6 +3937,7 @@ iseq_peephole_optimize(rb_iseq_t *iseq, LINK_ELEMENT *list, const int do_tailcal
 
                 if ((flag & VM_CALL_ARGS_SPLAT) && (flag & VM_CALL_KW_SPLAT) &&
                         (flag & VM_CALL_KW_SPLAT_MUT) && !(flag & VM_CALL_ARGS_BLOCKARG)) {
+write(1, "5", 1);
                     OPERAND_AT(iobj, 0) = Qfalse;
                 }
             }
@@ -3958,6 +3963,7 @@ iseq_peephole_optimize(rb_iseq_t *iseq, LINK_ELEMENT *list, const int do_tailcal
 
                     if ((flag & VM_CALL_ARGS_SPLAT) && (flag & VM_CALL_KW_SPLAT) &&
                             (flag & VM_CALL_KW_SPLAT_MUT) && (flag & VM_CALL_ARGS_BLOCKARG)) {
+write(1, "6", 1);
                         OPERAND_AT(iobj, 0) = Qfalse;
                     }
                 }
```

Program to load files and collect metrics (pipe the output of this into the next program):

```ruby
# coding: UTF-8
b = binding
ARGV.flat_map{|x| File.directory?(x) ? Dir["#{x}/**/*.rb"] : x}.each do |file|
  code = "# coding: UTF-8\nBEGIN{throw :valid, true}\n" + File.binread(file)
  catch(:valid){eval(code, b, file)}
rescue SyntaxError
end
puts
```

Program to parse metrics to get above output:

```ruby
res = $stdin.read

types = (<<END).split("\n")
f(1, *a)
f(1, *a, &lvar) | f(1, *a, &@ivar)
f(*a, **lvar) | f(*a, **@ivar)
f(*a, **lvar, &lvar) | f(*a, **@ivar, &@ivar)
f(*a, kw: 1)
f(*a, kw:1, &lvar) | f(*a, kw:1, &@ivar)
END

%w[1 2 3 4 5 6].each do |i|
  count = res.count(i)
  next unless count > 0
  print count.to_s.rjust(3), ' : ', types[i.to_i-1]
  puts
end
```