@byroot 2023-12-01
Have you tried this patch on `yjit-bench`? I suspect that if this translate to real world gains, it should be visible on things like the lobsters bench as it's a Rails app, so likely has significant delegation going on.