@jeremyevans 2023-12-01
> Have you tried this patch on `yjit-bench`? I suspect that if this translate to real world gains, it should be visible on things like the lobsters bench as it's a Rails app, so likely has significant delegation going on.

Here are the results.  I ran with `WARMUP_ITRS=1 MIN_BENCH_ITRS=2 MIN_BENCH_TIME=1` and it still took over 15 minutes on my machine.  For more accurate numbers, you would want to run it with standard settings on a machine that is not being used for anything except the benchmark.

Note that other than the compiler change to change `splatarray true` to `splatarray false` in certain cases (which should always improve runtime performance), the only other change in the pull request is the small one to vm_inshelper.c. That should not result in an additional allocation (pull request changes that allocation from caller-side to callee-side). Possibly combining the branches makes things slower, I could keep the split branches to remove the added conditional.

```
after: ruby 3.3.0dev (2023-11-30T23:30:30Z reduce-implicit-al.. b2707195ec) [x86_64-openbsd7.4]
before: ruby 3.3.0dev (2023-11-06T19:39:09Z v3_3_0_preview3~213 49b6dc8f07) [x86_64-openbsd7.4]

--------------  ----------  ----------  -----------  ----------  ------------  --------------
bench           after (ms)  stddev (%)  before (ms)  stddev (%)  after/before  before 1st itr
activerecord    123.9       3.1         122.3        0.4         1.01          1.04
erubi_rails     45.8        5.0         46.2         5.1         0.99          0.97
hexapdf         6239.6      0.1         6181.9       0.2         1.01          1.01
liquid-c        147.5       0.5         147.3        0.5         1.00          0.99
liquid-compile  173.0       0.3         177.5        0.6         0.97          0.98
liquid-render   318.4       0.5         320.2        0.1         0.99          1.00
mail            321.9       0.1         323.1        0.1         1.00          0.97
psych-load      5883.5      0.1         5790.1       0.2         1.02          1.02
railsbench      4092.9      0.5         4147.1       0.2         0.99          0.96
ruby-lsp        14.6        10.5        14.7         10.5        0.99          1.02
sequel          182.7       1.9         183.1        1.7         1.00          0.99
binarytrees     675.3       0.3         655.1        0.2         1.03          1.03
chunky_png      2207.5      0.0         2228.3       0.2         0.99          0.99
erubi           576.7       0.1         573.6        0.1         1.01          1.01
etanni          665.1       0.1         661.6        0.0         1.01          1.00
fannkuchredux   3741.2      0.2         3594.7       0.0         1.04          1.04
lee             2571.0      0.4         2550.8       0.3         1.01          1.01
nbody           168.7       0.2         169.0        0.2         1.00          0.99
optcarrot       10568.0     0.0         10501.7      0.3         1.01          1.00
ruby-json       10451.8     0.9         10318.8      0.0         1.01          1.00
rubykon         19591.2     0.1         19625.6      0.0         1.00          1.00
30k_ifelse      3390.4      0.0         3371.2       0.1         1.01          1.02
30k_methods     7287.3      0.0         7375.1       0.9         0.99          1.00
cfunc_itself    300.8       0.1         294.4        0.3         1.02          1.01
fib             439.6       0.1         432.4        0.3         1.02          1.02
getivar         204.1       0.1         213.7        0.2         0.95          0.96
keyword_args    506.8       0.2         491.4        0.0         1.03          1.02
respond_to      695.8       0.1         704.7        0.2         0.99          0.99
setivar         141.2       0.4         137.4        0.3         1.03          1.03
setivar_object  164.7       0.4         167.7        0.2         0.98          0.98
setivar_young   166.4       0.3         174.6        4.6         0.95          0.99
str_concat      186.0       2.0         185.5        2.1         1.00          1.01
throw           43.8        0.4         43.5         6.5         1.01          1.02
--------------  ----------  ----------  -----------  ----------  ------------  --------------
```